requirements.txt
torch>=1.7.0
torchvision>=0.8.1
numpy>=1.19.2
opencv-python>=4.4.0
Pillow>=8.0.0
matplotlib>=3.3.2
scikit-image>=0.17.2
tqdm>=4.50.2
tensorboard>=2.3.0